docker/app.py
from vetiver import VetiverModel
from dotenv import load_dotenv, find_dotenv
import vetiver
import pins

load_dotenv(find_dotenv())

b = pins.board_folder('/data/model', allow_pickle_read=True)
v = VetiverModel.from_pin(b, 'penguin_model', version = '20240307T071142Z-d45a2')

vetiver_api = vetiver.VetiverAPI(v)
api = vetiver_api.app
